细节优化

+++ src/components/TrashDetail.vue
@@ -59,6 +59,12 @@ export default {
     this.getTrashNotes()
       .then(() => {
         this.setCurTrashNote({ curTrashNoteId: this.$route.query.noteId })
+        this.$router.replace({
+          path: '/trash',
+          query: {
+            noteId: this.curTrashNote.id
+          }
+        })
       })
   },
 
@@ -93,11 +99,34 @@ export default {
     ]),
 
     onDelete () {
-      this.deleteTrashNote({ noteId: this.curTrashNote.id })
+      this.$confirm('删除后将无法恢复', '确定删除？', {
+        confirmButtonText: '确定',
+        cancelButtonText: '取消',
+        type: 'warning'
+      }).then(() => {
+        return this.deleteTrashNote({ noteId: this.curTrashNote.id })
+      }).then(() => {
+        this.setCurTrashNote({})
+        this.$router.replace({
+          path: '/trash',
+          query: {
+            noteId: this.curTrashNote.id
+          }
+        })
+      })
     },
 
     onRevert () {
       this.revertTrashNote({ noteId: this.curTrashNote.id })
+        .then(() => {
+          this.setCurTrashNote({})
+          this.$router.replace({
+            path: '/trash',
+            query: {
+              noteId: this.curTrashNote.id
+            }
+          })
+        })
     }
   }
